Revert "ci: run script to test bpftool types/options sync"
This reverts commit 03a098058467add9a1ae54cc92003a4afb2ce3b7.

The function was simply defined and not called in this commit. We tried
to call it in a later commit, but it would fail because the VM image
does not have python3 installed at this time.

We will use another approach and run the checks on the host directly,
not in the VM running the other selftest. Clean up the existing
function.