ci: ensure temp patch application is atomic
Fix the logic to do a dry-run of patching to not partially apply the
(ultimately) failing patch. This ensures that patch application is
all-or-nothin and won't be corrupting local source code.

While at it, clean up old temporary patches that are not necessary
anymore. Also add two patches necessary for bpf tree to run selftests
successfully.

Signed-off-by: Andrii Nakryiko <andrii@kernel.org>