ci: fix fold for the shutdown step
The marker for starting the "fold" section for the VM shutdown on the CI
interface is produced from a script running on the VM, and the tag was
created for Travis. Since then, most tags were replaced by a call to a
shell helper that produces the Travis or GitHub Action for the fold,
depending on the context. Let's update the marker for the fold for VM
shutdown too, to hide the related logs by default.

This would allow us to have a clearer output, and to add a more visible
summary of the tests, in a next step.