ci: fix test on exitstatus
The check on the $exitstatus collected from the /exitstatus file was
changed in a recent commit, and is not correct. Let's fix it. We
consider that we failed to read the exit status if we get something
other than an integer. This might happen if awk fails because it cannot
access the file, for example.

Let's also enquote the arguments for the "printf" near the end of the
script.

Fixes: d31757894acf ("ci: change VM's /exitstatus format to prepare it for several results")