Use libbpf/ci build-linux action
Instead of having a bunch of shell scripts here and a bunch more over in
libbpf/ci, let's consolidate everything in the latter. An earlier change
has already added an action containing the necessary logic to build the
Linux kernel image to that repository and with this change we make the
switch to using said action, allowing us to remove the build.sh and
helpers.sh scripts.

Signed-off-by: Daniel Müller <deso@posteo.net>