Run on self-hosted runners
We have seen cases of timeouts in the CI where a GitHub-hosted runner
was used for test execution and that slowed down everything from setup
to kernel build to running of tests.
Given that we have sufficient EC2 runners, which are beefier than
GitHub-hosted ones and, hence, can execute runs in a more timely fashion
and with fewer chances for timeouts, let's make sure to only use
self-hosted runners.
The self-hosted label is automatically assigned to all self-hosted
runners [0]. With this change we make sure to execute jobs only where
this label is present.

[0] https://docs.github.com/en/actions/using-workflows/workflow-syntax-for-github-actions#choosing-self-hosted-runners

Signed-off-by: Daniel Müller <deso@posteo.net>