Use deny/allow lists from upstream
So far we have relied on allow/deny lists maintained in this repository
to decide which tests to explicitly include/exclude from running in CI.
With recent changes [0] this information is now available in upstream
Linux.
As such, this change switches us over to using the upstream
configuration and removes the locally maintained one. This change is a
respin of fba2b3da955c8 ("Use deny/allow lists from upstream") in
combination with de16bf7a024df ("Add support for allow/deny list
additions") that keeps the existing denylist (but renames it) for the
case where the bpf tree is tested, which does not yet have the upstream
configuration.

[0] https://lore.kernel.org/bpf/165893461358.29339.11641967418379627671.git-patchwork-notify@kernel.org/T/#m2a97b0ea9ef0ddee7a53bbf7919e3f324b233937

Signed-off-by: Daniel Müller <deso@posteo.net>