Add patch fixing NULL pointer dereference
When using a more minimal configuration (as it landed upstream in the
bpf-next tree), we saw frequent NULL pointer dereferences in the kernel
causing VM crashes and CI run failures as a result.
The issue has been fixed [0] in a different repository and so for now we
just carry around the patch here.

[0] https://lore.kernel.org/bpf/165942025658.342061.12452378391879093249.stgit@devnote2/

Signed-off-by: Daniel Müller <deso@posteo.net>