Replace hardcoded LLVM version
Whenever the current LLVM version is bumped, we have to adjust the
version to use in CI, because snapshot packages of the previous version
may be removed from distribution repositories and can no longer be
installed. Furthermore, this adjustment has to happen across
repositories and there are dependencies between them, making the process
tricky and partly serial and the changes non-atomic.
With recent changes we have introduced an action that contains the LLVM
version to use in an action in the central libbpf/ci repository that
constitutes the source of truth.
With this change we make use of this action to determine the LLVM
toolchain with which to work.

Signed-off-by: Daniel Müller <deso@posteo.net>