Use pattern match for allow/deny list logic
We support excluding and including of tests in CI runs based on deny and
allow lists, respectively, maintained in separate files. Currently these
lists are fed to test_progs via the -b (and -t) argument. This argument
is evaluated as sub-string match, meaning that for the string 'align',
for example, it would prevent both 'align-8' and 'misaligned' from
running (if deny-listed). Most of the time that's not what we want: we'd
rather specify complete names. If the desire truly is for an entry to
match multiple items, that should be made more explicit via the usage of
pattern/wildcard syntax.
With this change we switch over to using -d/-a for passing deny and
allow information to the program, which treats arguments as exact
strings, but optionally allows for wildcard usage.

Signed-off-by: Daniel Müller <deso@posteo.net>