Generate job matrix using Python
Once we split the VM_Test job matrix into separate ones for building and
testing, we will require yet a more involved matrix. It is borderline
impossible to generate it in bash. Hence, with this change we switch
over to using Python. That has the added advantage that we no longer
need to double quote the contents and get some more type safety.

Signed-off-by: Daniel Müller <deso@posteo.net>