Temporarily pin libbpf/ci revision to use
We want to rename the travis-ci/ directory to the more generic ci/, as
we are no longer using Travis. When we earlier landed that change [0] we
saw issues because the libbpf/ci repository, which contains the majority
of the actions we use, still expects contents in travis-ci/.
To make the change atomic, this change temporarily pins the revision of
libbpf/ci that we use. Once libbpf/ci has been adjusted to work with the
ci/, we can switch this repository over atomically.

[0] https://github.com/kernel-patches/vmtest/pull/119

Signed-off-by: Daniel Müller <deso@posteo.net>