Run individual selftests in separate jobs
We finally have all the pieces in place to run move the running of
individual selftests into separate jobs, which allows said tests to be
run in parallel to each other as opposed to serially as was the case so
far.
This change makes adjusts our workflow to do exactly that. In order to
go about that, we generate a new test matrix that includes the various
tests, one for each configuration.

Signed-off-by: Daniel Müller <deso@posteo.net>