Name jobs more concisely & usefully
Our job names are no longer as explicit as they can be. For instance, it
is no longer trivial to see what are the build jobs vs. the test jobs.
There is also some redundant information included in the job names, such
as the kernel version ("LATEST") or the distribution ("ubuntu-latest").
Just including the architecture for which we build and the toolchain we
use should be sufficient and less at risk of being cut off in the UI.
For test jobs, lets additionally include the test we are about to run as
the first bit of information.

Signed-off-by: Daniel Müller <deso@posteo.net>