ci: Make run_selftests.sh read test names from boot parameters
When running the tests with qemu, we used to bake the name of the tests
to run as part of the root image.
this forces us to have each individual test_suite run build the image
(currently 4 times).
All of them are still copying all tests in the image though.

In order to avoid this duplication of work, we can build the image once
with all needed binaries, and then let the step running the tests pass
it as an argument.
The way to do that from outside the VM, is to leverage the boot
parameters and have some logic in the run_selftests.sh script to extract
those tests from /proc/cmdline.

Having a root image where we can instruct what test to run will also
allow us to essentially run the test on any platform able to emulate the
target architecture. For instance, we could run aarch64 tests on x86_64
host.

Another advantage is github aciton will pass this image as an artifact
from one step to another. This also means that the root image will be
readily available for download and we could easily reply the test on the
device of our choice.

The next step will be to change libbpf/ci actions to stop adding test in
the image and pass them when calling qemu.

Testing the parsing of run_tests from cmdline:
https://gist.github.com/chantra/5c7dd823e2b7c9419ef55d8139e82010

Testing sorting of 2 arrays:
https://gist.github.com/chantra/3c06ff7722568c928b79e3ba5b8ae5a7

Signed-off-by: Manu Bretelle <chantr4@gmail.com>