ci: Run test_progs with internal parallelism (-j) enabled
This patch is the continuation of @mykola-lysenko's work over in
https://github.com/kernel-patches/vmtest/pull/135 with the added feature
of not having failures of parallel selftests affect the overall run
results.
Note that GitHub Action's jobs.<job_id>.steps[*].continue-on-error [0]
feature works in such a way that you only see test failures in the
overall workflow summary and not even as individual job failures. That
means such failures are not communicated the most prominently, but it's
better than the other extreme of failing the entire run (which would
cause too much noise for BPF maintainers).

Here is a test run: https://github.com/kernel-patches/bpf/actions/runs/3324169845
It illustrates that 1) failures of serial tests (test_progs_no_alu32)
are still cause for overall workflow failure and 2) that failures of
parallel ones (e.g., test_progs_parallel) are not.

[0] https://docs.github.com/en/actions/using-workflows/workflow-syntax-for-github-actions#jobsjob_idstepscontinue-on-error

Closes: #149
Signed-off-by: Daniel Müller <deso@posteo.net>