ci: Include patches for detection of lockups & hung tasks
With this change we include two patches in CI: the first one is
available in bpf-next upstream, but has not yet made it to bpf. It
enables the kernel's panic on hard/soft lockup facility. The second one
has not been upstreamed yet (see commit description for reasoning). It
enables hung task detection as well, which is especially relevant now
that we are running test_progs with -j (which seems to trigger races
inside the kernel that cause hangs and, in turn, can lead to CI
timeouts). If there are no negative side effects of the latter, we can
upstream it later.

Signed-off-by: Daniel Müller <deso@posteo.net>