denylist: Add netcnt to denylist for x86_64
When ran under kvm, the test is failing with
`packets unexpected packets: actual 10001 != expected 10000`

this does not seem to reproduce under full qemu emulation.

In order to be able to introduce runners with kvm enabled, adding this
test to
the denylist will help keeping CI green while this is being
troubleshooted.

Signed-off-by: Manu Bretelle <chantr4@gmail.com>