workflow: re-arrange runner selection logic
If we are not in a vetted repo, only run on x86_64's ubuntu-latest
runners (GH runners)
Otherwise, run on runners that have both `self-hosted` and `uname -m`
labels.

This will allow to move self-hosted runners at org level and have them
being used across the different repos we support under kernel-patches
and make better use of the pool of runners.

Signed-off-by: Manu Bretelle <chantr4@gmail.com>