Gracefully fail parallel tests getting stuck
We have seen cases where the new parallel tests get stuck and we hit the
overall workload timeout (e.g., [0]). That is despite us having enabled
various panic on {soft lockup, hard lockup, task hang} settings already.
That can happen, for example, if test_progs itself (i.e., the user space
portion) encounters a deadlock.
The test being marked as failed in such a case is causing some noise in
Patchwork. To work around this issue, this change introduces a second
level of timeouts, specifically for parallel job runs. Because these
timeouts are now attached directly to the job, which in turn is marked
as continue-on-error, they no longer cause failure of the overall
workflow.

I have verified this behavior with the following GitHub Actions
configuration, in which the "Test" step hit the 1 minute timeout but the
overall job succeeded [1].

jobs:
  test:
    runs-on: ubuntu-latest
    timeout-minutes: 3
    steps:
      - name: Test
        continue-on-error: true
        timeout-minutes: 1
        shell: bash
        run: |
          sleep 120

[0] https://github.com/kernel-patches/vmtest/actions/runs/3371914924/jobs/5595049018
[1] https://github.com/danielocfb/all-things-test/actions/runs/3373709973/jobs/5598556394

Signed-off-by: Daniel Müller <deso@posteo.net>