Include base branch name in cache key identifier
It is arguably sub-optimal to share build artifacts between different
branches; bpf-next and bpf in our case. It works and is not technically
wrong, but it can result in unnecessary rebuilds, because divergence
between the two is much larger. So it makes sense to only use caches
from the actual base branch with priority.
With this change we adjust our cache key identifier logic accordingly.
The downside is that now more caches will be created and with the
addition of aarch64 we may actually run into the 10 GiB cache limit. We
will have to see whether that causes increased cache misses.

Signed-off-by: Daniel Müller <deso@posteo.net>