Work with relative times in incremental build logic
We are still seeing what appears to be excessive rebuilds in many CI
runs. One thing we did not consider in the initial version of the
incremental build logic is that different runners may have different
time stamps. While we do adjust source code time stamps, build artifact
time stamps are never touched -- but are ultimately influenced by the
system time.
With this change we adjust this logic to work with relative times
instead: we just set the source code time stamps to current time -2
minutes and the build artifact times to current time -1 minute. That is
really all we need and it makes everything independent of the system
time at the time of initial build. It also simplifies the working in
general, because less state is carried over between runs.

Signed-off-by: Daniel Müller <deso@posteo.net>