Work more with upstream commit SHA-1
So far our incremental build logic was remember the pull request/push
SHA-1 as the baseline for incremental builds. This commit information,
notably, already contains the CI files added by Kernel Patches Daemon.
That is conceptually fine, but it means that
1/ it's unnecessarily hard to reason about cache IDs, because they
   contained a commit SHA-1 that only made sense to the repository in
   question (it could still be fetched explicitly, but that's a
   conceptually unnecessary step)
2/ by default we have to perform an additional fetch from origin on the
   incremental build patch, because the SHA-1 we want to check out as
   the baseline is never present in our history; while typically this
   fetch is quick, we have seen it balloon to taking upwards of ten
   minutes (!!) at times; it's also just an unnecessary step
   conceptually

By using an upstream SHA-1 we circumvent these problems: the commit now
has a meaning in the bpf repository itself (1/) and most of the
time the SHA-1 will be present in the 50 commits we fetch from our
history (2/). The main downside is that there appears to exist no
reliable generic way to determine what is the most recent upstream
commit. So right now (on a push event) we just assume it is HEAD^. That
introduces a bit of a coupling to KPD, in that we assume that there is a
single commit stacked on top of upstream ("adding CI files"). That
should be acceptable, though, as nothing will break even if this
condition were to be violated.

Signed-off-by: Daniel Müller <deso@posteo.net>