Use more accessible time stamps in cache key
We currently use the commit date converted to Unix time stamp in our
cache key identifier. That works, but it is hard to read and reason
about.
This change switches to using ISO 8601 format instead. The wrinkle here
is that we want to use some kind of normalized time stamps, all lying in
the same time zone. With Unix time stamps a conversion to UTC was
happening automatically, but with ISO 8601 Git would simply encode the
committer's time zone in the resulting string, without a conversion
happening. That can subtly screw up lookups of the most recent cache. To
fix that, use local ISO 8601 format and override the local time zone to
be UTC.

Signed-off-by: Daniel Müller <deso@posteo.net>